refactor: move colorizeVersionDiff to colorizeUtils.ts
